+++ components/checks/active/code_injection.rb
@@ -12,8 +12,6 @@
 #
 # @author Tasos "Zapotek" Laskos <tasos.laskos@arachni-scanner.com>
 #
-# @version 0.2.3
-#
 # @see http://cwe.mitre.org/data/definitions/94.html
 # @see http://php.net/manual/en/function.eval.php
 # @see http://perldoc.perl.org/functions/eval.html
@@ -32,8 +30,8 @@ class Arachni::Checks::CodeInjection < Arachni::Check::Base
 
     def self.options
         @options ||= {
-            substring: (rand1.to_i + rand2.to_i).to_s,
-            format:    [Format::STRAIGHT]
+            signatures: (rand1.to_i + rand2.to_i).to_s,
+            format:     [Format::STRAIGHT]
         }
     end
 
@@ -70,7 +68,7 @@ Injects code snippets and assess whether or not execution was successful.
 },
             elements:    ELEMENTS_WITH_INPUTS,
             author:      'Tasos "Zapotek" Laskos <tasos.laskos@arachni-scanner.com>',
-            version:     '0.2.3',
+            version:     '0.2.4',
             platforms:   payloads.keys,
 
             issue:       {